// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ChannelVisibility = {
  ALL: 'ALL',
  REGISTERED_USERS: 'REGISTERED_USERS',
  ALLOWLIST: 'ALLOWLIST',
  PASSWORD: 'PASSWORD',
  UNLISTED: 'UNLISTED'
} as const

export type ChannelVisibility = (typeof ChannelVisibility)[keyof typeof ChannelVisibility]


export const RecordingVisibility = {
  PUBLIC: 'PUBLIC',
  ALLOWLIST: 'ALLOWLIST',
  UNLISTED: 'UNLISTED',
  PRIVATE: 'PRIVATE'
} as const

export type RecordingVisibility = (typeof RecordingVisibility)[keyof typeof RecordingVisibility]


export const RecordingQueueState = {
  UPLOADING: 'UPLOADING',
  UPLOADED: 'UPLOADED',
  PENDING: 'PENDING',
  ENCODING: 'ENCODING',
  ENCODED: 'ENCODED',
  MERGING: 'MERGING',
  MERGED: 'MERGED',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  EXPIRED: 'EXPIRED'
} as const

export type RecordingQueueState = (typeof RecordingQueueState)[keyof typeof RecordingQueueState]